+++ src/protocol/qbsprotocolhelloresponse.ts
@@ -4,9 +4,11 @@ import { QbsProtocolDataKey } from './qbsprotocoldatakey';
 export class QbsProtocolHelloResponse {
     public readonly apiLevel: number = 0;
     public readonly apiCompatibilityLevel: number = 0;
+    public readonly lspSocket?: string;
 
     public constructor(response: any) {
         this.apiLevel = parseInt(response[QbsProtocolDataKey.ApiLevel]);
         this.apiCompatibilityLevel = parseInt(response[QbsProtocolDataKey.ApiCompatLevel]);
+        this.lspSocket = response[QbsProtocolDataKey.LspSocket];
     }
 }